+++ package.json
@@ -3,11 +3,13 @@
   "private": true,
   "version": "1.0.0",
   "scripts": {
-    "generate": "node ./generate.js"
+    "generate": "node --trace-deprecation ./generate.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "quicktype-core": "^23.0.170"
+    "fetch": "^1.1.0",
+    "quicktype-core": "^23.0.170",
+    "typescript": "^5.6.3"
   }
 }
